+++ src/Export.jl
@@ -7,7 +7,7 @@ import Pluto: ServerSession
 using HTTP
 using Base64
 using SHA
-using Sockets
+import Pkg
 
 
 myhash = base64encode ∘ sha256
@@ -219,6 +219,40 @@ function try_tocache(cache_dir::String, hash::String, state)
     end
 end
 try_tocache(cache_dir::Nothing, hash, state) = nothing
+
+
+
+## FINDING THE PLUTO VERSION
+
+
+function try_get_pluto_version()
+    try
+        deps = Pkg.API.dependencies()
+
+        p_index = findfirst(p -> p.name == "Pluto", deps)
+        p = deps[p_index]
+
+        if p.is_tracking_registry
+            p.version
+        elseif p.is_tracking_path
+            error("Do not add the Pluto dependency as a local path, but by specifying its VERSION or an exact COMMIT SHA.")
+        else
+            # ugh
+            is_probably_a_commit_thing = all(in(('0':'9') ∪ ('a':'f')), p.git_revision)
+            if !is_probably_a_commit_thing
+                error("Do not add the Pluto dependency by specifying its BRANCH, but by specifying its VERSION or an exact COMMIT SHA.")
+            end
+
+            p.git_revision
+        end
+    catch e
+        @error "Failed to get exact Pluto version from dependency. Your website is not guaranteed to work forever." exception=(e, catch_backtrace())
+        Pluto.PLUTO_VERSION
+    end
+end
+
+
+
 
 
 end
